Added option to view USED files
Now has the ability to show both used and unused files so you can validate that it's keeping/deleting the right files. You can't pick and choose from there--that might be an option in the future, but for now you'll just end up with a list that you'll need to delete (or restore) manually.
Also fixed a few bugs, so that it gets closer to 'right'.